Move validation file
